Make server into an SCGI application
This is an experiment. It currently handles only one request at a time,
which is a step backwards compared to CGI (which would start a new
process for each request): if this works, it should be modified to spawn
a thread/process for each request.

The hope is that this approach will allow long-running concurrent
processing (e.g., refresh()) to happen while still serving the request
quickly. (Errors during refresh would need to be logged elsewhere,
though, i.e., not via cgi_die.)