Workers track status and can restart tasks.
The worker programs will now record what they're doing for any
given job in status/job_<id>

If a task finishes, and the last line in status/job_<id> is
'Restart' then the worker will reattempt the task. This allows
an admin to restart a task that has encountered a transient
deadlock.