Filter out PRs with test-failing labels.
Recognise some labels as requests not to test the relevant PR. These
include the "test failing" label currently being used for PRs that have
failed a regression test. I've also added a "no test" label, which doesn't
yet exist in github, that could be applied to PRs that require discussion
but not a regression test.