Fix server dying on sending Zulip message
The server code prior to this commit appears to expect a 204 No
Content response when successfully sending a Zulip message, but the
response Zulip gives is often (always?) a 200 OK instead. Thus, even
when the Zulip messages are sent correctly, the server interprets it
as a failure, and propagates this failure to the worker. The worker
then aborts, since instead of receiving the expected Finished response
to the API call that triggered the Zulip message, it instead receives
something like this:

Error:
Error sending Zulip message
{"result":"success","msg":"","id":521754098}
200

...where the last two lines are what the server received from
Zulip.

I have not tested this commit beyond compiling it, since I do not run
a regression server; I hope @xrchz can do this.

For the same reason, the generalised "204 or 202" check is written in
a way that is arguably a bit generous in what it accepts.

One could also argue that there's no reason for the server to crash
the worker because sending a Zulip message failed, since this is not a
worker concern at all. If so, there's an even simpler fix: just remove
the assert entirely.